CP-2275:  Remove feature flagged direct-to-Cognito signup code (#4463)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2275

## 🛠 Changes

- Remove obsolete Create Account and Verify Account code, now that we have all sign up requests going through the API, rather than direct to Cognito

## ℹ️ Context for reviewers

We previously had a decent bit of complexity in our sign up and account verification logic because the API user would be created through the Cognito Post-Confirmation Lambda trigger, which would be triggered _after_ an account is verified. That meant we had to pass metadata, like whether the user is claimant/employer and an EIN # into the account verification request. We no longer need to do this, and those Lambda triggers are getting removed as well.

I have one follow-up PR to also cleanup some obsolete code in the Reset Password flow.