CP-2275: Remove UserNotFound workaround on Password Reset page (#4464)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2275

## 🛠 Changes

- Remove obsolete Reset Password code, now that we have all sign up requests going through the API, rather than direct to Cognito

## ℹ️ Context for reviewers

We previously had a workaround in our password reset logic to address an issue where our Cognito Post Confirmation Lambda trigger sometimes didn't create the API user record. For this instance, we'd redirect a user to the Reset Password page, because resetting a password re-triggers the Post Confirmation trigger, which would re-attempt creating the API user record. 

We no longer need this workaround, because the API user record is now created during the initial sign up request, and if it fails, the user would never be able to login. We'll also be removing the Post Confirmation lambda soon.