CP-2227: Update maintenance page times (#4469)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2227

## 🛠 Changes

* Updates the July 13 maintenance page start and end dates per new requirements

## ℹ️ Context for reviewers

This is a prod hotfix, so just this commit will get promoted through stage and prod. I tested the new dates locally by applying the changes to my development environment and adjusting my computer's date and time. This will be UATed in stage before going to prod.

The pagerduty maintenance window has been updated

## 👀 Demo

<img width="457" alt="Screen Shot 2021-06-13 at 3 00 17 AM" src="https://user-images.githubusercontent.com/9039672/121695925-8976bf80-ca99-11eb-8a3a-03e0ea7d4bcd.png">

## 🌡 End to End testing

N/A