PFMLPB-1436: modify E2E test to cover future leave payments (#4451)
Modifications to the existing medical_reduced_approval_approval_payment_30k.ts and the scenario used for that test.

Created FINEOS helper function to assert correct payment dates.

It has been brought to our attention that we currently do not have any coverage for future leaves. We have decided that it would be best to modify the leave dates on one of our existing tests in order to test future leaves. A payment test was modified in order to verify that future, paid leave cases are being created with the correct information (i.e payment dates, payment amounts).

* modify test to cover future leave payments

* rename test file and update correct scenario