EMPLOYER-1438: Use singular concurrent_leave (#4496)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1438

## 🛠 Changes

Change `concurrent_leaves` (an array of 0 or 1 elements) to `concurrent_leave` (an optional/nullable concurrent leave object) in the employer claim POST request. This will match the employer claim GET request interface and also the interface used in the applications API.

## ℹ️ Context for reviewers

There was some confusion/lack of alignment on whether to use `concurrent_leave` or `concurrent_leaves` and we ended up using both in different places. This PR brings the various applications / claims APIs into alignment on `concurrent_leave` since there will only ever be 0 or 1 concurrent leave.

## 👀 Demo

N/A

## 🌡 End to End testing

N/A