INFRA-521: Disable API ECS Task start-failure notifications (#4579)
https://lwd.atlassian.net/browse/INFRA-521

    Ignore the API ECS Task failure notifications for starting.

ℹ️ Context for reviewers

Occasionally, an API container will fail to start due to AWS networking issues. This is normal and not a cause for alarm because generally a new container will replace it successfully. This only happens during deployments so we would also know if there are issues.

Additionally, when AWS has long-standing issues or our VPC begins to have networking issues, the API will start to rotate tasks pretty frequently and the notifications will spam slack =)