EMPLOYER-1410: Store checkbox field values as arrays (#4592)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1410

## 🛠 Changes

- Update `useHandleInputChange` hook so that it stores checkbox field values as an array of checked options

## ℹ️ Context for reviewers

This is to support the Leave Admin dashboard's "Filter by status" UI, which will allow a user to select multiple statuses. Each input has `claim_status` as the `name`.

I did an audit of all our usages of `InputChoiceGroup` and all usages are a radio group. We haven't had any checkbox fields in the app up until this point.