EMPLOYER-1382: Clear cached EmployerClaim after info request is submitted (#4628)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1382

## 🛠 Changes

- Fixes issue where the cached `EmployerClaim` included `is_reviewable: true` even after it was reviewed
- Updates dashboard links to use client-side navigation now that the above if addressed

## ℹ️ Context for reviewers

`EmployerClaim` is populated by the `GET` request, which makes a handful of calls to FINEOS to get that data. The `PATCH` request avoids making those calls, and therefore doesn't pass back any data besides the absence case ID, which means the Portal doesn't have data to update the cached claim.