INFRA-504 Remove Temporary Manage Step Functions API Deployment Step (#4673)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-504

🛠 Changes
Removed the check-step-function-management job
ℹ️ Context for reviewers
In #4218 the management for step functions moved from the api modules to the ecs-task modules. As a cleanup measure, there was a temporary step that destroyed the targeted modules if they were being managed by the api. This removed a circular dependency when the ECS tasks were built.