INFRA-564 Check for PUB Process File Everyday (#4700)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-564

🛠 Changes
Created a scheduled CI action that will check for a file in the s3://massgov-pfml-prod-agency-transfer/reports/sent/ bucket
ℹ️ Context for reviewers
On the night of June 22nd, the pub-payments-process-fineos ECS task failed to start due to an AWS issue. We received a PagerDuty warning and the issue was resolved in the morning, but it took some time for the PUB team to become aware of the issue.

The PUB team would like to have a critical page go to the API on-call rotation in the morning if the task fails or does not complete for any reason.