INFRA-577: Ignore FINEOSFatalUnavailable errors explicitly in New Relic transactions (#4746)
https://lwd.atlassian.net/browse/INFRA-577

    Ignore FINEOSFatalUnavailable errors explicitly in New Relic

ℹ️ Context for reviewers

I noticed that these errors are still showing up as TransactionErrors and in the Errors Inbox. #3516 should have ignored them based on the returned status code, but I suspect because the FINEOS errors are subclasses of RuntimeError instead of werkzeug HTTPException, they are not recognized as 503s.

This is my hunch because PaymentRequired inherits from HTTPException, specifies "402" on the class, and is properly ignored by the ignore_status_codes: 402 statement in the newrelic.ini.

Instead of going down the rabbit hole right now to verify and fix this behavior for FINEOS exceptions I'm just going to add the exception class directly as a class to ignore.