CP-2372: Remove flakey synthetic login check (#4793)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-2372

## 🛠 Changes

- Remove the Portal's synthetic login monitor due to issues keeping it from being blocked by Cognito.

## ℹ️ Context for reviewers

Cognito identifies login attempts from a range of IP addresses as suspicious, and will begin blocking login attempts from that account. A user needs to reset their password in order to get unblocked. The account being used for the synthetic monitor was getting blocked, resulting in the monitor failing and a low priority alarm being triggered. In lieu of figuring out a way to not have the account get blocked, I'm removing the synthetic monitor. 

**How will we know if login isn't working then?**

We will still have an alarm policy that triggers when there is a high number of failed login requests:

https://github.com/EOLWD/pfml/blob/df411a81108c830675302e8a90dc635f1ca99e66/infra/modules/alarms_portal/alarms-newrelic.tf#L192-L200