INFRA-450: Update API error rate alarm and add minimum user thresholds (#4763)
https://lwd.atlassian.net/browse/INFRA-450
🛠 Changes

    Add minimum user thresholds of 10 users to API error rate and network error rate alarms
    Remove filters for errors that have been fixed
    Add filters for errors that are currently being triaged

ℹ️ Context for reviewers

We've been receiving more error rate alarms recently, several of which have woken up on-call engineers unnecessarily in the middle of the night. I've pulled this out of the backlog due to the noise frequency.

The Portal error rate alarms have thresholds for the minimum number of unique sessions needed in order for the alarm to trigger. This was implemented in #3478. This PR implements the same for the API error rate alarms.

I've chosen 10 users kinda arbitrarily, but think it's a kinda reasonable number. We have pretty high usage during the day so in a lot of cases this won't affect our alarm results.