END: Do not allow unstable E2E hangs to trigger GHA job failure (#4813)
https://lwd.atlassian.net/browse/PFMLPB-1587

Addresses an issue where unstable tests were hanging endlessly, causing the Github Job to time out, and failure notifications to be triggered, even though the "failure" was happening in the unstable test suite, which we don't alert on. We work around this issue by adding a step timeout to the stable and unstable test steps, forcing the step to time out before the job timeout is reached. When the step timeout happens, continue-on-error is respected, meaning it will ultimately prevent any unstable test hang from triggering a job failure (and hence, a notification).