INFRA-536 Vacate the Release Namespace in API Makefile (#4757)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-536

🛠 Changes
rename some of the commands in api/Makefile
ℹ️ Context for reviewers
None of the below tasks participate in the business process of creating releases; instead, they implement or describe technical parts of the deployment process: building, tagging, and pushing either the API's Docker image, or else a local Docker cluster composed from the API image and an ad-hoc database image. Room needs to be made in this namespace, so these technical tasks don't become confused with several soon-to-be-written business tasks.

In the API Makefile, we have [release, release-build, release-publish, release-start-api, release-stop-api, release-test, release-publish, release-push, release-tag] and the variable RELEASE_TAG. All these tasks need renaming to emphasize their role in building or publishing the API's Docker image.