PFMLB-1620: Create a template for Intake through FINEOS (#4840)
🎫 Ticket
https://lwd.atlassian.net/browse/PFMLB-1641
Subtask: https://lwd.atlassian.net/browse/PFMLB-1620

🛠 Changes
This PR will add a template to allow us to create claims through Fineos intake process. The Fineos will create an Absence Case that allows us to adjudicate the claim if needed. This will help and speed up testing for CPS testing.

ℹ️ Context for reviewers
This is ignored folder so it can only be run manually. It will allow us to create claims in Fineos through the intake process and create an Absence Case.