END: clean up submitClaimToAPI types (#4879)
Changes the types related to the submitClaimToAPI Cypress task to reflect the reality of underlying code, so that there's no need to make some assertions inside our test specs. I.e. we don't need to assert on existence of fineos_absence_id property, because the API submission code already does that for us.