END: Move payment preference checking into Page objects (#4883)
Moves checking of the payment details (Banking info, address) out of fineos.ts and into fineos.pages.ts. Note that there is a minor functional change here, which is that we're now checking the payment details on the claim rather than the employee. This seems more correct to me (and is faster to do, since we're in the claim pages anyway).

The test this is actually used in is in the ignored folder though, so we won't see any actual change right now - it just moves code around and makes it easy in the future to check payment details.