EMPLOYER-1542: Use query param as source of truth for filter and search state (#4930)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1542

## 🛠 Changes

- Remove dependency on the `activeFilters` prop being set on the `Dashboard` page, and utilize the values set in the URL query string instead.

## ℹ️ Context for reviewers

I've split this out of #4831 so that it can be easier to review. The reason we need to not rely on the `activeFilters` prop is because it is set by the `withClaims` higher-order component (HOC). In #4831, the `Dashboard` page will no longer be wrapped by that HOC, so therefore will no longer have access to the `activeFilters` prop. We instead can get the same information from the query string, which is where those components store those values.

## 👀 Demo

This PR should have no user-facing impacts.