END: Refactor extension notification test to be more performant (#4986)
This PR refactors the extension notification (approval) test to be more stable/performant when the test is flaky. For an example of how this test is slow during retries, see this run, which took 5 minutes for the test due to retries on 2/3 steps.

What we did:

Split claim submission/approval and extension of benefits into two separate steps. This prevents a retry on either step from duplicating work that takes several minutes to complete.
Limit retries on the notification step, and extend the wait time for the notification e-mail so we never execute this long wait twice.
Also, when I was updating this test, I encountered the "Out of date data" error, which seemed to be resolved by manually triggering the legal notice after initial submission/approval.