END: Block New Relic script using global beforeEach (#4992)
Cypress and New Relic don't play well together, and we've always been blocking the NR script from loading, but have recently learned that we're not blocking the script in all cases. This PR resolves this issue by moving the NR script interception to a global beforeEach() block. This should prevent spurious New Relic error logs in all of our tests.