END: Add debug statement to watch URLs as they change (#4993)
Adds a commented-out debug statement. When enabled, this code will output each URL the Cypress browser switches to. Hopefully, this will help us debug future test hangs resulting from hitting Fineos error pages that we aren't modifying the content of. We're leaving this snippet disabled as it produces a lot of log data that probably isn't necessary in most cases.