API-1740: PI Claims Status phase 1.0 - /claims enhancement Mockup (#4969)
Created a DetailedClaimResponse with leave period and outstanding evidence information. Defined this new model both in the openapi.yaml file and the response models. Modified the /claims/{fineos_absence_id} endpoint to use this response.

Separately added an AbsencePeriodFactory to aid in the unit test setup.

This ticket covers only the modeling of the new objects not the retrieval of data. The LeavePeriodStatusResponse model has a from_orm method defined for now to validate the model works using the database in the unit test. In the future we may remove this method and fill in the model with FINEOS API data.