END: Reduce timeouts for notification feature tests (#5024)
Reduces timeouts for some of the notification/notice checks during our feature tests to avoid long-running tests when notifications are not working properly. Our goal here is to reduce the impact of a "worst case scenario" where notifications do not arrive. In practice, most of our notifications are actually arriving pretty quickly.