API-1784 - Create an eligibility feed for each employer in input list. (#5031)
Added an entry point, and method, to receive a list of employer identifiers for which we need to produce a feed of all their employees.

Logic to gather all employees is the same as that for the ALL mode. First use case for this mode is to fix Lowe's and UU LLC. The list of employers will never be too long. Lowe's has about 9K employees.