PFMLPB-1731: Remove trigger-cypress action (#5045)
https://lwd.atlassian.net/browse/PFMLPB-1731

In #5034, we added concurrency control to the e2e-cypress and e2e-flood workflow that's used to run E2E tests and load tests. In this PR, we stop using the bespoke trigger-cypress action. This action was being used because it allowed us to detect a pre-existing E2E run for any given commit. With concurrency being handled per-environment now, there's no more reason we need to limit concurrency ourselves, so using a third party action is now preferable.

The other substantive change here is that we're now using the workflow (and test code) from the main branch. Since these jobs were originally written, we've switched to using a backward compatible approach to E2E testing (E2E tests should be able to run against recent versions of the portal/API successfully), so it is more advantageous for us to use test code from main when testing releases, as that code is more likely to be stable than the release branch, which we don't have control over.

The result has been tested with a Portal deploy of main to test: https://github.com/EOLWD/pfml/runs/3304046658?check_suite_focus=true. I tried an API deploy (also main to test), but it was unsuccessful for other reasons unrelated to this work (https://github.com/EOLWD/pfml/runs/3304029531?check_suite_focus=true).