EMPLOYER-1633: Move validation files from services/ to validation/ (#5072)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1633

## 🛠 Changes

Moved our validation-related files (i.e `application_rules.py`) out of `services/` and into `validation/` instead, so that validation-related code is in a common directory.

## ℹ️ Context for reviewers

These files, like `application_rules.py`, include methods for running validations and returning any validation errors. Because of that, I think these make more sense in the `validation/` directory, which can serve as the common location for validation-related code.

This is one of likely several PRs to refactor validation code in the API codebase.