EMPLOYER-1638: Clear cached claims after info request is reviewed (#5095)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1638

## 🛠 Changes

- Clear the cached set of claims displayed on a Leave Admin's dashboard after they review a claim, so that the claim's status displayed on the dashboard is accurate.

## ℹ️ Context for reviewers

A user might click the "Back to Dashboard" button on the success page, after they've reviewed a claim. In this scenario, without this change, they'd land on the dashboard and it would say that the claim they just reviewed still requires an action. That's not accurate. We want the dashboard to instead indicate that no action is required.