PFMLPB-1827: Pass environment variables to Artillery container (#5136)
https://lwd.atlassian.net/browse/PFMLPB-1827

Changes the way we get environment variables into our containers. Instead of compiling .env into the Docker container or into the code, we pass only the ones we need at runtime to AWS. In local development, the values are pulled in from the .env file in the root.

Note: The fact that the .env file was compiled into the code was a legacy thing from when I was trying to run in Artillery cloud. We shouldn't need to compile it anymore, now that we have control over container environment variables.