PFMLPB-1878: Reinstate New Relic reporter with less test hangs (#5177)
https://lwd.atlassian.net/browse/PFMLPB-1878

Reinstates the New Relic Cypress reporter introduced in #5163 and removed in #5175. This reporter was causing test hangs, which I was able to trace back to an unhandledRejection handler in Mocha. It looks like one limitation of reporters is that they absolutely cannot throw an unhandled promise rejection. This PR changes the reporter so it logs and swallows any fetch errors, since this functionality is not mission-critical.

In the process of debugging, I removed cypress-multi-reporters, choosing to instead extend the New Relic reporter from the standard spec reporter. So now we still get nice console output as well as NR reporting, but don't need a third party module to do it.