PFMLPB-1760: Refactor medical_intermittent_notice_notification_denial.ts to include checking the status in the Claimant portal (#5185)
🛠 Changes

    Update medical_intermittent_notice_notification_denial.ts for coverage on the claimant claim status feature.
    Add portal actions to support testing new feature

ℹ️ Context for reviewers

The claim status page is currently using static data which cannot be used for meaningful coverage. This didn’t block the development of the test however, we will have to come back and update our assertions once claim statuses use live data if we merge before that time. (9/7 according to this ticket)