PFMLPB-1761: Refactor bonding_reduced_notice_notification_approval.ts to include checking the claim status in the claimant portal (#5201)
🎫 Ticket

https://lwd.atlassian.net/browse/PFMLPB-1769
Subtask: https://lwd.atlassian.net/browse/PFMLPB-1761
🛠 Changes

    Update bonding_reduced_notice_notification_approval.ts for coverage on the claimant claim status feature.

ℹ️ Context for reviewers

Currently, the claimant status feature is deployed to test and stage in addition to using mock data on the claim status page. There are a few to-do items commented in the files that will need to be done once this new feature is deployed to all environments/stops using mock data.