END: Reduce video compression to improve test speed (#5206)
🛠 Changes

Reduces video compression for the videos Cypress captures to speed up E2E runs. The default setting for compression is 32 CRF, and this changes it to 20. This has an impact of speeding up the overall run by spending less time in compression. The downside is that videos files are larger, which increases the storage space on Cypress.io, but since we aren't billed on this metric, it doesn't really matter to us.