API-1783: Claims Status phase 1.2 - route for single absence case (#5137)
Modified /claims/:fineos_absence_id to use DetailedClaimResponse, which was defined in a previous PR. This response includes the application_id associated to the claim as well as the absence periods for the absence case in FINEOS.

Added a mock response to mock_client for unit tests. During development I tested with FINEOS dt2 and perf environments.