PFMLPB-1784: Split integration tests into stable/unstable/nightly (#5233)
🎫 Ticket

https://lwd.atlassian.net/browse/PFMLPB-1784

🛠 Changes

Splits our integration tests into stable/unstable/nightly groups that are triggered with the same logic as Cypress tests. The motivation for this was to offload performance-draining work from running in the test environment during the day, where we're already seeing instability. My hope is that moving the leave admin push into a nightly test will improve stability of this environment overall.