END: Pull config values out of LST deployment code (#5184)
We'd previously added code that used hard-coded values pointing to AWS resources. This is not ideal, since it's inflexible, and we'd prefer not to have these committed. This PR removes the hard-coded values and moves that configuration into a "parameter" in the AWS account we're using for deployment. This means anyone with access to the account has access to those parameters (note: these values are not sensitive).

This PR also refactors the LST deployment code in the following ways:

Use a class to manage the ECS deployment flow rather than procedural code. This helps us contain the use of the config values and handle the complexity of fetching the config from AWS.
When running docker commands in the script, use integrate'd I/O, so we get a full log of what Docker is doing in the background.
Update from aws-sdk to @aws-sdk packages (v2 -> v3). This wasn't strictly necessary, but since I was updating some of the commands it seemed easy to do. I also updated the other place where we use this code (the integration test).