INFRA-681 Reconfigure Load Balancer Settings (#5246)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-681

🛠 Changes
Changed the values of healthy_threshold and unhealthy_threshold
ℹ️ Context for reviewers
healthy/unhealthy thresholds represent the number of times the health check needs to succeed (or fail) before marking the instance as healthy or unhealthy. Since the default interval time is 30 seconds, that means that:

Lower environments (with ~3 containers) have less time to become healthy before they are marked as unhealthy and are replaced.

Upper environments (with 10 containers) require at least 5 minutes of healthy activity before they are marked healthy and the deployment is completed.