EMPLOYER-1405: Updated New Relic page view tracking logic (#5171)
## 🎫 Ticket

https://lwd.atlassian.net/browse/EMPLOYER-1405

## 🛠 Changes

* Updated logic for new relic interactions
* Setting Attributes when we know them
* Handling edge cases where duplicate tracking previously occurred / no tracking at all

## ℹ️ Context for reviewers
Attributes were previously being tracked as soon as the page start began. Setting attributes when we know if the user is logged in or not will prevent the tracker from firing before the data is available. 

Due to how NextJS route changes work, I decided to break up our tracking into a few separate parts. 

**Not logged in**
When a user is not logged in, visiting an unauthenticated page, these page tracks will be recorded in it's own `useEffect`, outside of the routing `useEffect`. This is done because the route change does not take account for a `hard refresh`. This is what is causing the issue where for example, the homepage at `/` does not get tracked. 

**Logged in**
This functions primarily the same as the previous implementation, except the logic has been moved to `handleRouteChangeComplete` && a check that the user is not on the `/applications` page. See comment in Files change for more context.

**Logged in, but initial redirect**
In the case where a user is logged in and visits the page again, they will be redirected to `/applications` initially. (If the user is an employer, they get redirected to `/applications` then `/employers/welcome/`. This causing inaccurate reporting for the employer side, since we shouldn't track `/applications` for the employers.