END: Partition generated credentials to avoid duplicate e-mail error (#5330)
Updates the credentials we generate for claimants so that they are "partitioned." During a recent load test, we ran into a rash of "this e-mail is already registered" errors. Adding a new part to the tag of the e-mail address we generate will allow us to work around this issue. In the future, this partition number can be incremented to ensure a bunch more unique addresses.