remove unnecessary code from spec (#5396)
Removes code checking for adding DOB and Address to claimant info in fineos from fineos_medical_continous_reduction_denial.ts. It was necessary when this test was checking for payments, as they can't be generated without a DOB and address. Now this test is denying the claim and payments are not generated.