fix to approval test (#5419)
🎫 Ticket
n/a

🛠 Changes
This PR addresses recent errors when clicking on "View status updates and details". Sometimes when clicked there would be a wait and the test would try to assert data on the wrong page (see here). This PR adds another assertion after the click to make sure you're on the status page.