PFMLPB-1983: Add synthetic check for Fineos that collects version string (#5417)
Adds a new New Relic synthetics check for Fineos. This check runs every 10 minutes and in addition to telling us whether Fineos is up or not, also scrapes the Fineos version string found on the bottom of the UI. This data is collected as part of the synthetics event, where it is queryable. Ultimately, we hope to use this to keep track of Fineos deployments and react to them accordingly.

This check is only being added to the lower environments without SSO right now, because SSO brings an additional layer of complexity (handling SSO login), and the version strings in those environments are currently hard-coded for an unknown reason.