PFMLPB-1622 Create an automated test to support testing of CPS work pattern through API causing incorrect weighting days (#5502)
🎫 Ticket

https://lwd.atlassian.net/browse/PFMLPB-1622
🛠 Changes

This was to test to make sure that we get the correct weight days in Fineos when a new claimant creates a claim in the middle of the week. This was fixed in the June release, but regressed in the July release and was fixed in the August release. We submitting a new claim with no hours against the claimant. This claimant submits a claim through the Portal for Wednesday, not Sunday. We check in Fineos to see the weight days match the other days (0.20) and it not showing 1.00 instead. Note this can't be used for E2E Suite because we need to make sure the claimant is unused to run this test correctly.
ℹ️ Context for reviewers

To test use the claimant who does not have a birthday set in env. Add a directory to the data folder e2e/data/cps-testing/employees.json and e2e/data/cps-testing/employers.json in those JSON files add the EE/ER. Update the config.json file to use these directories as well. Run the following testing the ignored folder e2e/cypress/specs/ignored/cps-sp/continous_medical_midwk_work_patterns_weight_days.ts