API-1792: Eligibility Feed export should only create one employer file with all pending employee updates (#5315)
Removed the batching logic from the updates mode and the batch_size configuration parameter. Adjusted the test cases accordingly.

Added a test - test_process_employee_updates_for_multiple_employers_produce_equal_number_of_files - to ensure one file per employer is created, regardless of order of data in WagesAndContributions table.

The Eligibility Feed export code in “updates” mode grabs EmployeeLog records to process in batches (50k in all environments at time of writing). This was to potentially support certain forms of parallel processing that has never materialized.

If there are more EmployeeLog records than the batch size, it’s possible all the employees for a given employer will not be in the same batch and so we’ll end up generating multiple Eligibility Feed files for the same employer.

As the files are timestamped with the generation time, the files don’t conflict, but it’d be cleaner if all employees to be exported for a given employer are grouped together.