PUB-291: Add tables to contain the raw data from MMARS for payments (#5554)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PUB-291

## 🛠 Changes

Added tables to store the raw MMARS data for payments. The full schema we expect from MMARS can be seen on https://lwd.atlassian.net/wiki/spaces/DD/pages/1924825131/Backfilling+MMARS+Payments

## ℹ️ Context for reviewers

As a first step in reconciliation, taxes, 1099 work and a whole host of other work, we first need to store the raw version of MMARS payment data in our DB to work with.

A follow-up PR will come shortly that shows how we'll use this to make actual payment records, which will use the payment_id column to attach the created payment.

The payments table corresponds to paid payments, the refunds table corresponds to overpayment recoveries.