END: Fix failing nightly Cypress test group (#5658)
🎫 Ticket

No ticket.
🛠 Changes

Fixes an issue where the nightly run group in Cypress is failing in Github actions. This comments out the nightly run group - we can re-enable it when we actually have nightly tests to run.

Also resolves a linting issue where actionlint was throwing an error on e2e-cypress.yml due to our use of the new run_attempt context property. This issue has been raised upstream, so this is a temporary workaround: rhysd/actionlint#57

* END: Add missing values causing nightly tests to error out

* Adjust actionlint to ignore new property

* Add ignore to actionlint

* Fix syntax

* Another syntax change