docs: refresh INFRA README documentation (#5372)
The infra README has gotten kinda bloated and a good chunk of it isn't relevant for folks visiting the directory for the first time.

Additionally, the team has noticed some gaps in documentation and shared knowledge around how we do certain common things, like creating new ECS tasks; testing infra changes in a live environment; and testing Github Action changes.

The files are numbered so they appear in a sensible order, both in Github and in the rendered docs website. My intent is to begin organizing documentation to be more of a "guided" learning book, instead of a hodgepodge of random knowledge strewn amongst different files and directories.