PAY-12: quick fix to create db migration to fix typos in indexes (#5694)
🎫 Ticket
No ticket, but related to https://lwd.atlassian.net/browse/PAY-12
#5566

🛠 Changes
Added a db migration. The original indexes were created wiht a typo ...1009... and subsequently were renamed in the model to be ...1099... and also change some constraints, but the migration file was not included. Needed here to have these updates separate of other unrelated alembic migrations. Followup to some changes from the original ticket #5566

There is also another change being generated by alembic. This is likely due to some alembic-internal feature that reads these indexes differently when they are added as part of the model.

    op.drop_constraint(
        "ix_employee_occupation_employee_id_employer_id", "employee_occupation", type_="unique"
    )
    op.create_index(
        "ix_employee_occupation_employee_id_employer_id",
        "employee_occupation",
        ["employee_id", "employer_id"],
        unique=True,
    )
In ...employees.py#1090, the index is added in the model class, whereas we more commonly add these manually in the alembic migration file. There may be a difference in how alembic interprets this, so auto-generating this new migration file also drops/recreates the same effective index/constraint.