PFMPB-1871: E2E test that will assert withdrawn status in claimant portal and download the withdrawn notice (#5704)
Refactor fineos_medical_continuous_notice_historical_withdraw.ts so that it is now submitting a caring leave claim. The test additionally now logs into the claimant portal and verifies the claim status of withdrawn as well as downloading the withdrawn legal notice.

Currently this functionality is only ready to be tested in DT2, with the release to other environments coming later. The HAS_WITHDRAWN_STATUS variable has been added to config.json to only run this spec in environments that support the withdrawn status.