API-1835: automated employee demographics update from DUA CSV (#5400)
🎫 Ticket
https://lwd.atlassian.net/browse/API-1835

Changes/Context
Setting up a runnable script for a one-off one-time execution to backfill gender values from a DUA CSV. The test verifies the logic of writing gender values only if there is a null value already. This accounts for if DUA sends multiple rows for the same claimant, or if we somehow fill the gender value in between sending the request file to DUA and receiving the response.

This script will also ignore invalid gender values, i.e. DUA will only send M, F or U(nknown) which map to our expected values.

Demo
Ran this locally by running make login and then manually running the poetry script. The test file is the same one that the unit test uses, and is tested against a local database that has two existing entries (fineos_id, gender)